% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.check.R
\name{conv.check}
\alias{conv.check}
\title{Convergence diagnostics on fitted model output.}
\usage{
conv.check(object, also.unpen = FALSE)
}
\arguments{
\item{object}{Fitted model object.}

\item{also.unpen}{If \code{TRUE}, displays eigenvalues range also for the unpenalised Hessian. Defaults to \code{FALSE}.}
}
\value{
Convergence diagnostics.
}
\description{
Convergence diagnostics on fitted model output.
}
