% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.fmsm.R
\name{logLik.fmsm}
\alias{logLik.fmsm}
\title{Extract the log likelihood for the fitted multistate model}
\usage{
\method{logLik}{fmsm}(object, ...)
}
\arguments{
\item{object}{Fitted model object of class \code{fmsm} produced by function \code{\link{fmsm}}.}

\item{...}{Unused in this case.}
}
\value{
Standard logLik object.
}
\description{
It extracts the log-likelihood for a fitted \code{fmsm} model.
}
