% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fmsm.R
\name{plot.fmsm}
\alias{plot.fmsm}
\title{Function to plot the smooths included in the model specifications.}
\usage{
\method{plot}{fmsm}(x, ...)
}
\arguments{
\item{x}{Fitted model object.}

\item{...}{Other graphical arguments.}
}
\value{
Plots the smooths.
}
\description{
Function to plot the smooths included in the model specifications.
}
