% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fmsm.R
\name{print.fmsm}
\alias{print.fmsm}
\title{Print a fmsm object}
\usage{
\method{print}{fmsm}(x, ...)
}
\arguments{
\item{x}{\code{fmsm} object produced by function \code{\link{fmsm}}.}

\item{...}{Unused in this case.}
}
\value{
\code{print.fmsm} prints out a matrix summarising the positions of the transition intensities, the transition intensities formulae, the total number of observations, etc for the fitted multistate survival model.
}
\description{
The print method for the \code{fmsmObject} produced by \code{\link{fmsm}}.
}
