% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.fMmsm.R
\name{print.summary.fMmsm}
\alias{print.summary.fMmsm}
\title{Flexible transition intensity based models for univariate multistate processes}
\usage{
\method{print}{summary.fMmsm}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{Fitted model object.}

\item{digits}{Number of digits printed in the output.}

\item{signif.stars}{By default significance stars are printed alongside the output.}

\item{...}{Other arguments.}
}
\value{
Prints model term summaries.
}
\description{
Flexible transition intensity based models for univariate multistate processes
}
