% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateIDM.R
\name{simulateIDM}
\alias{simulateIDM}
\title{Function to predict and plot the estimated transition intensities (and confidence intervals).}
\usage{
simulateIDM(N = N, seed = seed, og.12 = TRUE)
}
\arguments{
\item{N}{Total number of individuals.}

\item{seed}{Seed used for the simulation.}

\item{og.12}{If \code{TRUE} a common shape for the first transition is used. If \code{FALSE} a sinusoid is used.}
}
\value{
Simulated data generated from an Illness-Death model (IDM).
}
\description{
Function to predict and plot the estimated transition intensities (and confidence intervals).
}
