% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fMmsm.R
\name{summary.fMmsm}
\alias{summary.fMmsm}
\title{Summary for fitted model ouput.}
\usage{
\method{summary}{fMmsm}(object, ...)
}
\arguments{
\item{object}{Fitted model object.}

\item{...}{Other arguments.}
}
\value{
Summary of fitted model object.
}
\description{
Summary for fitted model ouput.
}
